requirements.txt
discord.py
python-dotenv
libsql-client
pytz
apscheduler
